backend-recommend/requirements.txt
asgiref==3.6.0
automium==0.2.6
automium-web==0.1.1
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.6
Django==4.2
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
Flask==2.2.3
gensim==4.3.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
konlpy==0.6.0
legacy==0.1.6
lxml==4.9.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
py4j==0.10.9.5
PyMySQL==1.0.3
pytz==2023.3
regex==2023.3.23
scipy==1.10.1
smart-open==6.3.0
sqlparse==0.4.3
tqdm==4.65.0
tzdata==2023.3
Werkzeug==2.2.3
zipp==3.15.0
